backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic==2.7.1
google-generativeai==0.7.2
pandas==2.2.2
scikit-learn==1.5.0
numpy==1.26.4
python-multipart==0.0.9
playwright==1.44.0
playwright-stealth==1.0.6
